+++ ctdcal/processors/functions_oxy.py
@@ -1,6 +1,8 @@
 """
 Oxygen functions for processing sensor data, unit conversions and derived variables.
 """
+from collections import namedtuple
+
 import gsw
 import numpy as np
 
@@ -253,3 +255,260 @@ def calculate_dV_dt(oxy_volts, time, nan_replace=True):
     # filtered_dvdt = scipy.signal.filtfilt(b, a, dv_dt)
 
     return dV_dt  # filtered_dvdt
+
+
+# Rinko functions
+RinkoO2Cal = namedtuple("RinkoO2Cal", [*"ABCDEFGH"])
+RinkoTMPCal = namedtuple("RinkoTMPCal", [*"ABCD"])
+
+
+def rinko_DO(p_prime, G, H):
+    """
+    Calculates the dissolved oxygen percentage.
+    """
+
+    DO = G + H * p_prime
+
+    return DO
+
+
+def rinko_p_prime(N, t, A, B, C, D, E, F, G, H):
+    """
+    Per RinkoIII manual: 'The film sensing the water is affect by environment
+    temperature and pressure at the depth where it is deployed. Based on experiments,
+    an empirical algorithm as following is used to correct data dissolved oxygen.'
+
+    Parameters
+    ----------
+    N : array-like
+        Raw instrument output
+    t : array-like
+        Temperature [degC]
+    A-H : float
+        Calibration parameters
+    """
+    p_prime = A / (1 + D * (t - 25)) + B / ((N - F) * (1 + D * (t - 25)) + C + F)
+
+    return p_prime
+
+
+def correct_pressure(P, d, E):
+    """
+    Parameters
+    ----------
+    P : array-like
+        Temperature-corrected DO [%]
+    d : array-like
+        Pressure [MPa]
+    E : float
+        Manufacturer calibration coefficient
+
+    Returns
+    -------
+    P_d : array-like
+        Temperature- and pressure-corrected DO [%]
+    """
+    # what is the dbar ~ MPa?
+
+    P_d = P * (1 + E * d)
+
+    return P_d
+
+
+def salinity_correction(DO_c, T, S):
+    """
+    Oxygen optode is not able to detect salinity, so a correction is applied to
+    account for the effect of salt on oxygen concentration. See Uchida (2010) in
+    GO-SHIP manual (pg. 6, eq. 9) for more info.
+
+    Parameters
+    ----------
+    DO_c : array-like
+        Pressure-corrected dissolved oxygen
+    T : array-like
+        Calibrated CTD temperature
+    S : array-like
+        Calibrated CTD salinity
+
+    Returns
+    -------
+    DO_sc : array-like
+        Pressure- and salinity-corrected dissolved oxygen
+    """
+    # solubility coefficients from Benson and Krause (1984),
+    # as recommended by Garcia and Gordon (1992)
+    B0 = -6.24523e-3
+    B1 = -7.37614e-3
+    B2 = -1.03410e-2
+    B3 = -8.17083e-3
+    C0 = -4.88682e-7
+
+    # "scaled temperature"
+    T_scaled = np.log((298.15 - T) / (273.15 + T))
+
+    # correction equation
+    DO_sc = DO_c * np.exp(
+        S * (B0 + (B1 * T_scaled) + (B2 * T_scaled ** 2) + (B3 * T_scaled ** 3))
+        + C0 * S ** 2
+    )
+
+    return DO_sc
+
+
+def _Uchida_DO_eq(coefs, inputs):
+    """
+    See Uchida et. al (2008) for more info:
+    https://doi.org/10.1175/2008JTECHO549.1
+    and Uchida et. al (2010) - GO-SHIP manual
+
+    Parameters
+    ----------
+    coefs : tuple
+        (c0, c1, c2, d0, d1, d2, cp)
+    inputs : tuple
+        (raw voltage, pressure, temperature, salinity, oxygen solubility)
+    """
+    c0, c1, c2, d0, d1, d2, cp = coefs
+    V_r, P, T, S, o2_sol = inputs
+
+    K_sv = c0 + (c1 * T) + (c2 * T ** 2)  # Stern-Volmer constant (Tengberg et al. 2006)
+    V0 = (1 + d0 * T)  # voltage at zero oxygen (Uchida 2010, eq. 10)
+    Vc = (d1 + d2 * V_r)  # raw voltage (Uchida 2010, eq. 10)
+    o2_sat = ((V0 / Vc) - 1) / K_sv  # oxygen saturation [%] (Uchida 2010, eq. 6)
+
+    DO = o2_sat * o2_sol  # dissolved oxygen concentration
+    DO_c = DO * (1 + cp * P / 1000) ** (1 / 3)  # pressure compensated DO
+    DO_sc = salinity_correction(DO_c, T, S)  # salinity + pressure compensated DO
+
+    return DO_sc
+
+
+def oxy_weighted_residual(coefs, weights, inputs, refoxy, L_norm=2):
+    """
+    A weighted residual fit in a similar manner to that of the SBE43 method of oxy_fitting.py.
+    """
+    # (abstracted from PMEL code oxygen_cal_ml.m)
+    # unweighted L2: sum((ref - oxy)^2)  # if weighted fails
+    # unweighted L4: sum((ref - oxy)^4)  # unsure of use case
+    # unweighted L1: sum(abs(ref - oxy))  # very far from ideal
+    # anything else? genericize with integer "norm" function input?
+
+    residuals = np.sum(
+        (weights * (refoxy - _Uchida_DO_eq(coefs, inputs)) ** 2)
+    ) / np.sum(weights ** 2)
+
+    return residuals
+
+
+def rinko_temperature(v, tmp_cal:RinkoTMPCal):
+    """
+    Calculate rinko temperature from voltage and calibration coeffieicnets.
+    """
+    if type(tmp_cal) is not RinkoTMPCal:
+        raise ValueError("tmp_cal must be of type RinkoTMPCal")
+
+    A, B, C, D = tmp_cal
+    return A + B*v + C*v**2 + D*v**3
+
+def rinko_pprime_aro_cav(v, t, o2_cal:RinkoO2Cal):
+    """
+    Calculates Rinko P' of the equation P = G + H * P'
+    where P is DO physical value IN PERCENT [%]
+    """
+    A, B, C, D, E, F, G, H = o2_cal
+
+    term_1_denominator = 1 + D*(t-25) + F*(t-25)**2
+    term_1 = A/term_1_denominator
+
+    term_2_denominator = v * (1 + D*(t-25) + F*(t-25)**2) + C
+    term_2 = B/term_2_denominator
+
+    return term_1 + term_2
+
+def rinko_saturation(pprime, o2_cal:RinkoO2Cal):
+    """
+        Calculates Rinko P of the equation P = G + H * P'
+        where P is DO physical value IN PERCENT [%]
+    """
+
+    A, B, C, D, E, F, G, H = o2_cal
+
+    return G + H * pprime
+
+def rinko_correct_for_pressure(p, d, o2_cal:RinkoO2Cal):
+    """Note that the pressure term, d, must be in MPa
+
+    1 decibar = 0.01 Mpa
+    """
+    A, B, C, D, E, F, G, H = o2_cal
+
+    return p*(1 + E*d)
+
+#   20240612 AJM commenting out method, possibly redundant
+# def rinko_saturation(df, film="B", model="ARO-CAV", **kwargs):
+#     """
+#     Derive oxygen saturation for the RINKO. Unused as of 2021.
+#     """
+#     pass
+
+def rinko_oxy_eq(press, temp, oxyvo, os, o2_cal:RinkoO2Cal):
+    """
+    Derive corrected RINKO dissolved oxygen.
+    """
+    #Calculate pprime
+
+    pprime = rinko_pprime_aro_cav(oxyvo,temp,o2_cal)
+
+    # Calculate P (DO physical value in %)
+
+    #p = rinko_saturation(pprime, o2_cal)
+
+    # Correct for pressure * d is pressure in Mpa *
+
+    d = press * 0.01
+
+    p_corr = rinko_correct_for_pressure(pprime,d,o2_cal)
+
+    # Divide by 100 to get percents in the form of 0.xx
+
+    p_corr = p_corr / 100
+
+    # Multiply by OS to get DO (os can be in either ml/l or umol/kg)
+
+    DO = p_corr * os
+
+    return DO
+
+def rinko_curve_fit_eq(X, a, b, c, d, e, f, g, h):
+    """
+    Same as rinko_oxy_eq, but in a form that is more suitible for scipy's curve fit routine
+    X contains pressure, temperature, voltage, and OS (the normal arguments for rinko_oxy_eq)
+    """
+
+    press, temp, oxyvo, os = X
+    o2_cal = RinkoO2Cal(a, b, c, d, e, f, g, h)
+
+    #Calculate pprime
+
+    pprime = rinko_pprime_aro_cav(oxyvo,temp,o2_cal)
+
+    # Calculate P (DO physical value in %)
+
+    #p = rinko_saturation(pprime, o2_cal)
+
+    # Correct for pressure * d is pressure in Mpa *
+
+    d = press * 0.01
+
+    p_corr = rinko_correct_for_pressure(pprime,d,o2_cal)
+
+    # Divide by 100 to get percents in the form of 0.xx
+
+    p_corr = p_corr / 100
+
+    # Multiply by OS to get DO (os can be in either ml/l or umol/kg)
+
+    DO = p_corr * os
+
+    return DO
+
